Rename mcjf_sdformat to mjcf_to_sdformat
