+++ .github/workflows/ci.yml
@@ -34,11 +34,11 @@ jobs:
       run: |
         cd sdformat_mjcf
         python3 -m tox -e py
-    - name: Test mjcf_sdformat
+    - name: Test mjcf_to_sdformat
       env:
         PYTHONPATH: ${{ env.pythonLocation }}/lib/python${{ matrix.python }}/site-packages:/usr/lib/python3/dist-packages
       run: |
-        cd mjcf_sdformat
+        cd mjcf_to_sdformat
         python3 -m tox -e py
     - name: Upload Coverage to Codecov
       uses: codecov/codecov-action@v2
