+++ mjcf_to_sdformat/setup.cfg
